Refactor: 데이터 로직 커스텀훅으로 분리
